backend/requirements.txt
fastapi 
uvicorn 
motor 
passlib[bcrypt] 
python-jose 
python-dotenv
python-multipart